Adds changes for generating register allocated 3AC
This CR adds the code for generating register allocated 3AC using the
`RegisterFile` construct that we defined in the previous CR. In addition
to the changes for generating the register allocated 3AC, in this CR I
had to address some missed gaps in the `RegisterFile` type, the
major one being handling spilling of temporary variables. The way I'm
handling spills of temporaries is by creating associated
`FunctionScopedSymbols` for them when we do need to spill them. Handling
spilling of temporaries in this way should simplify the story of
generating tiny code for spills/stores of temporaries to stack memory.

Some of the methods in the methods in the `RegisterFile` type need to be
improved as they are reaching into the guts of the type where we can
introduce a layer of abstraction using additional methods. For instance,
in the `ensure_register` method I am directly accessing the
`tempi_to_func_scoped_symbol_map` field as a hashmap, thus making it
hard for my future self to refactor things if required.

I also still need to add unit tests for the `RegisterFile` type.

+++ src/symbol_table/scope.rs
@@ -11,16 +11,16 @@ use std::cell::RefCell;
 use std::fmt::{Display, Formatter};
 use std::panic::panic_any;
 use std::rc::Rc;
-use std::sync::atomic::{AtomicI32, AtomicU32, Ordering};
+use std::sync::atomic::{AtomicI32, AtomicU32, Ordering, AtomicUsize};
 
-static STACK_FRAME_LOCAL_SLOT_COUNTER: AtomicU32 = AtomicU32::new(1);
-static STACK_FRAME_PARAM_SLOT_COUNTER: AtomicU32 = AtomicU32::new(1);
+static STACK_FRAME_LOCAL_SLOT_COUNTER: AtomicUsize = AtomicUsize::new(1);
+static STACK_FRAME_PARAM_SLOT_COUNTER: AtomicUsize = AtomicUsize::new(1);
 
 pub fn reset_stack_frame_local_slot_counter() {
     STACK_FRAME_LOCAL_SLOT_COUNTER.store(1, Ordering::SeqCst);
 }
 
-pub fn get_stack_frame_local_slot_counter() -> u32 {
+pub fn get_stack_frame_local_slot_counter() -> usize {
     STACK_FRAME_LOCAL_SLOT_COUNTER.fetch_add(1, Ordering::SeqCst)
 }
 
@@ -28,7 +28,7 @@ pub fn reset_stack_frame_param_slot_counter() {
     STACK_FRAME_PARAM_SLOT_COUNTER.store(1, Ordering::SeqCst);
 }
 
-pub fn get_stack_frame_param_slot_counter() -> u32 {
+pub fn get_stack_frame_param_slot_counter() -> usize {
     STACK_FRAME_PARAM_SLOT_COUNTER.fetch_add(1, Ordering::SeqCst)
 }
